Moved ACL param creation test util function to a separate file. Updated comment on ACL test.
